pyproject.toml
[project]
name = "alterego"
version = "2.5.0"
description = "AlterEgo — a styled Discord bot for game nights, announcements, ads, events, and more"
requires-python = ">=3.12"
dependencies = [
    "discord.py==2.4.0",
    "aiohttp==3.11.18",
    "better-profanity==0.7.0",
    "flask==3.1.1",
]
